+++ src/traits.jl
@@ -7,8 +7,8 @@
 
 type matrixTopologicalOrder
     V::Matrix # Matrix in itself
-    nodesNumbers::Vector{Int64} # Vector of nodes numbers for ordering of the matrix
-    tipsNumbers::Vector{Int64} # Tips numbers
+    nodesNumbers::Vector{Int} # Vector of nodes numbers for ordering of the matrix
+    tipsNumbers::Vector{Int} # Tips numbers
     tipNames::Vector # Tip Names
     indexation::AbstractString # Are rows ("r"), columns ("c") or both ("b") indexed by nodes numbers in the matrix ?
 end
